Week4/backend/requirements.txt
Flask>=2.3,<3
flask-swagger-ui>=4.11.1,<5
PyJWT>=2.8,<3
